Add two styling classes
I noticed an issue first on mobile. The signature editor was quite wonky. I found that if you add the classes that are present for the About Me editor ("bio-composer" and "input-xxlarge"), the CSS handles the rest and it styles things much nicer on both mobile and desktop. Let me know if you need any extra info!